+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - TableView bound to ListProperty should rebind when value changes
 - Allow calling Workspace.disableNavigation() when workspace is empty
 - Thread pools are reinitialized on App.start() to support stop/start multiple times within the same JVM
- 
+- ServiceLoader provided interceptors were called every time an App class was initialized 
 - inheritParamHolder and inheritScopeHolder are cleared on Application stop (https://github.com/edvin/tornadofx/issues/602)
 - smartResize throws exception for hidden columns (https://github.com/edvin/tornadofx/issues/606)